Setting Up an R-Package
There are several ways for setting up and R-Package, in this course we are going to create a project in R-Studio for creating a basic package. For this it is necessary to install the next R packages:
devtools
roxygen

And to follow the first step that is shown in 'First_Step.pdf'.